+++ Campaign/app/Http/Controllers/BannerController.php
@@ -10,6 +10,7 @@ use App\Http\Showtime\Showtime;
 use App\Models\Dimension\Map as DimensionMap;
 use App\Models\Position\Map as PositionMap;
 use App\Models\Alignment\Map as AlignmentMap;
+use App\Variable;
 use Carbon\Carbon;
 use HTML;
 use Illuminate\Http\Request;
@@ -103,6 +104,7 @@ class BannerController extends Controller
             'positions' => $this->positionMap->positions(),
             'dimensions' => $this->dimensionMap->dimensions(),
             'alignments' => $this->alignmentMap->alignments(),
+            'variables' => Variable::all()->pluck('value', 'name'),
         ]);
     }
 
@@ -186,6 +188,7 @@ class BannerController extends Controller
                 'positions' => $this->positionMap->positions(),
                 'dimensions' => $this->dimensionMap->dimensions(),
                 'alignments' => $this->alignmentMap->alignments(),
+                'variables' => Variable::all()->pluck('value', 'name'),
             ]),
             'json' => new BannerResource($banner),
         ]);
@@ -208,6 +211,7 @@ class BannerController extends Controller
             'positions' => $this->positionMap->positions(),
             'dimensions' => $this->dimensionMap->dimensions(),
             'alignments' => $this->alignmentMap->alignments(),
+            'variables' => Variable::all()->pluck('value', 'name'),
         ]);
     }