+++ Campaign/app/Http/Requests/VariableRequest.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\Http\Requests;
+
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rule;
+
+class VariableRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    public function rules()
+    {
+        return [
+            'name' => [
+                "required",
+                Rule::unique("variables", "name")->ignore($this->variable->id ?? null)
+            ],
+            'value' => 'required',
+        ];
+    }
+}