+++ Campaign/app/Http/Showtime/ShowtimeResponse.php
@@ -12,7 +12,7 @@ interface ShowtimeResponse
 
     public function success(string $callback, $data, $activeCampaigns, $providerData);
 
-    public function renderBanner(Banner $banner, array $alignments, array $dimensions, array $positions): string;
+    public function renderBanner(Banner $banner, array $alignments, array $dimensions, array $positions, array $variables): string;
 
     /**
      * render is responsible for rendering JS to be executed on client.
@@ -24,5 +24,5 @@ interface ShowtimeResponse
      * @param $positions
      * @return string
      */
-    public function renderCampaign(CampaignBanner $variant, Campaign $campaign, array $alignments, array $dimensions, array $positions): string;
+    public function renderCampaign(CampaignBanner $variant, Campaign $campaign, array $alignments, array $dimensions, array $positions, array $variables): string;
 }