+++ Campaign/app/Observers/Variable.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Observers;
+
+class Variable
+{
+    public function saved()
+    {
+        \App\Variable::refreshVariableCache();
+    }
+}