+++ Campaign/resources/assets/js/components/BannerForm.vue
@@ -175,7 +175,7 @@
                                     <label for="target_url" class="fg-label">Target URL</label>
                                     <input v-model="targetUrl" class="form-control fg-input" name="target_url" type="text" id="target_url">
                                 </div>
-                                <div><small>RTM (REMP's UTM) params will be automatically appended to this link.</small></div>
+                                <div class="help-block"><small>RTM (REMP's UTM) params will be automatically appended to this link.</small></div>
                             </div>
 
                             <div class="input-group fg-float m-t-30">
@@ -187,7 +187,8 @@
                                         Object <i>params</i> contains several properties of the banner you can access.
                                         <span data-toggle="tooltip"
                                               data-original-title="properties: rtmSource, rtmMedium, rtmCampaign, rtmContent, rtmVariant"
-                                              class="glyphicon glyphicon-question-sign"></span>
+                                              class="glyphicon glyphicon-question-sign"></span><br>
+                                        <span v-pre>You can use <i class="zmdi zmdi-code"></i> Variables in this field as <code>{{&nbsp;variable_name&nbsp;}}</code>.</span>
                                     </small>
                                 </div>
                             </div><!-- .input-group -->
@@ -198,7 +199,8 @@
                                     <label for="jsIncludesStr" class="fg-label">Javascript includes</label>
                                     <textarea v-model="jsIncludesStr" class="form-control fg-input" rows="6" name="jsIncludesStr" cols="50" id="jsIncludesStr"></textarea>
                                 </div>
-                                <div style="margin-top: 5px;">Enter include script urls separated by new line.</div>
+                                <div style="margin-top: 5px;">Enter include script urls separated by new line. <br></div>
+                                <small class="help-block" v-pre>You can use <i class="zmdi zmdi-code"></i> Variables in this field as <code>{{&nbsp;variable_name&nbsp;}}</code>.</small>
 
                                 <div v-if="jsIncludes">
                                     <input v-for="jsInclude in jsIncludes" type="hidden" name="js_includes[]" :value="jsInclude">
@@ -213,6 +215,7 @@
                                     <textarea v-model="cssIncludesStr" class="form-control fg-input" rows="6" name="cssIncludesStr" cols="50" id="cssIncludesStr"></textarea>
                                 </div>
                                 <div style="margin-top: 5px;">Enter include stylesheets urls separated by new line.</div>
+                                <small class="help-block" v-pre>You can use <i class="zmdi zmdi-code"></i> Variables in this field as <code>{{&nbsp;variable_name&nbsp;}}</code>.</small>
 
                                 <div v-if="cssIncludes">
                                     <input v-for="cssInclude in cssIncludes" type="hidden" name="css_includes[]" :value="cssInclude">
@@ -228,7 +231,7 @@
                                     <small class="help-block">
                                         Banner events (show, click, close) are tracked back to Campaign tool automatically.
                                         Enable if you need to manually control when events are fired.
-                                        To track events manually, use <i>remplib.tracker.trackEvent()</i> function.
+                                        To track events manually, use <code>remplib.tracker.trackEvent()</code> function.
                                         See <a href="https://github.com/remp2020/remp/tree/master/Beam#js-tracking-interface">documentation</a> for further details.
                                     </small>
                                 </label>
@@ -392,6 +395,7 @@
                                         :positionOptions="positionOptions"
                                         :show="show"
                                         :template="template"
+                                        :variables="variables"
 
                                         :mediumRectangleTemplate="mediumRectangleTemplate"
                                         :overlayRectangleTemplate="overlayRectangleTemplate"
@@ -478,6 +482,7 @@
         "_alignmentOptions": Object,
         "_dimensionOptions": Object,
         "_positionOptions": Object,
+        "_variables": Object,
 
         "_validateUrl": String,
         "_clientSiteUrl": String,
@@ -545,6 +550,7 @@
             alignmentOptions: [],
             dimensionOptions: [],
             positionOptions: [],
+            variables: {},
             show: true,
 
             submitAction: null,
@@ -565,7 +571,7 @@
             cssIncludes: null,
             manualEventsTracking: null,
 
-            fieldParamsMessage: "RTM (REMP's UTM) params will be automatically appended to every link in this field.<br> If you want to add custom parameter to specific link: add data-param-* attribute. e.g.: data-param-foo=\"baz\""
+            fieldParamsMessage: "RTM (REMP's UTM) params will be automatically appended to every link in this field.<br> If you want to add custom parameter to specific link: add <code>data-param-*</code> attribute. e.g.: <code>data-param-foo=\"baz\"</code>"
         }),
         computed: {
             isOverlay: function() {