+++ Campaign/resources/views/banners/_form.blade.php
@@ -16,6 +16,7 @@
     var alignments = JSON.parse('{!! json_encode($alignments) !!}');
     var dimensions = JSON.parse('{!! json_encode($dimensions) !!}');
     var positions = JSON.parse('{!! json_encode($positions) !!}');
+    var variables = {!! json_encode($variables) !!};
 
     var banner = remplib.banner.fromModel({!! $banner->toJson() !!});
     banner.show = true;
@@ -23,6 +24,7 @@
     banner.dimensionOptions = dimensions;
     banner.positionOptions = positions;
     banner.forcedPosition = 'absolute';
+    banner.variables = variables;
     banner.validateUrl = {!! @json(route('banners.validateForm')) !!};
     banner.clientSiteUrl = '{{ Config::get('app.client_site_url') }}';