+++ Campaign/resources/views/layouts/app.blade.php
@@ -120,6 +120,9 @@
             <li {!! route_active(['campaigns']) !!}>
                 <a href="{{ route('campaigns.index') }}" ><i class="zmdi zmdi-ticket-star"></i> Campaigns</a>
             </li>
+            <li {!! route_active(['variables']) !!}>
+                <a href="{{ route('variables.index') }}" ><i class="zmdi zmdi-code"></i> Variables</a>
+            </li>
         </ul>
     </aside>